test(onboard): split --agents lifecycle into focused legacy-command-agents.test.ts
